[WSCoordinateAnalysis] Change primeVertical algorithm
The algorithm in reference documents for computing the
prime vertical is:

    \frac{a}{\sqrt{1 - e^2 \sin^2 \phi }}

where \phi is latitude.

i.e.
    a/sqrt(1 - e**2 * sin(latitude)**2)

This commit changes the algorithm to

    \frac{b}{\sqrt{1 - e^2 \cos^2 \phi }}

i.e.
    b/sqrt(1 - e**2 * cos(latitude)**2)

The table below represents the old and new
algorithms, comparing the length of the output
to the expected distance:
    old         new
    +00.36%     +00.12%
    +13.69%     +13.32%
    +00.37%     +00.12%
    +00.68%     +00.43%
    +00.90%     +00.63%

The table shows that the new algorithm is closer
to the expected distance for all trials.